pyproject.toml
[project]
name = "flowpower-core"
version = "0.1.0"
license = { text = "Business Source License 1.1" }
description = "Lean runner + tracer for PromptFlow"
authors = [{ name = "Matías" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer",
    "rich",
    "promptflow",
    "pyyaml"
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
fp = "flowpower.cli:cli"